+++ src/components/ui/project/jigsawpuzzle/PixiRejectWrapper.tsx
