+++ src/components/ui/project/jigsawpuzzle/pixi/AnimatedGIF.tsx
