+++ src/components/ui/project/jigsawpuzzle/pixi/CreditsRenderer.tsx
