+++ src/components/ui/project/jigsawpuzzle/pixi/PieceDisplay.tsx
