+++ src/components/ui/project/jigsawpuzzle/pixi/Preview.tsx
