+++ src/components/ui/project/jigsawpuzzle/pixi/PuzzleCompleteModal.tsx
