+++ src/components/ui/project/jigsawpuzzle/pixi/SmoothGraphics.tsx
