+++ src/components/ui/project/jigsawpuzzle/puzzle/Piece.tsx
@@ -5,7 +5,7 @@ import React, {
 } from 'react';
 import { Container, Sprite } from '@pixi/react';
 import { Container as PixiContainer, DisplayObject, Texture } from 'pixi.js';
-import { PUZZLE_WIDTH } from '@/components/ui/project/kroniipuzzle/puzzle/PuzzleConfig';
+import { PUZZLE_WIDTH } from '@/components/ui/project/jigsawpuzzle/puzzle/PuzzleConfig';
 import PuzzleStoreContext from '../providers/PuzzleStoreContext';
 import Message from './Message';
 import PieceInfo from './PieceInfo';
