+++ src/components/ui/project/jigsawpuzzle/puzzle/PieceInfo.ts
